+++ RELEASE_NOTES.md
@@ -13,6 +13,7 @@
  * [Add OHLC Chart]()
  * [Add option to render all charts derived from the scatter trace type via WebGL as scattergl trace]()
  * [Add Waterfall Chart]()
+ * [Add ScatterGeo, PointGeo, LineGeo Charts]
 
 ### 1.2.2 - Apr 9 2020
  * [Opening Charts is now more or less OS agnostic](https://github.com/muehlhaus/FSharp.Plotly/commit/f6e3dceade085e43e7e56b478b9cf7b533a4fe55)